move live tab functions to tabs.js
