+++ js/tabs.js
@@ -1,5 +1,5 @@
-import { updateSummaryPollutantChart, updatepm10LineChart } from "./chart.js";
-import { fetchWAQIData } from "./api.js";
+import { updateSummaryPollutantChart, updatepm10LineChart, updatePollutantLiveChart } from "./chart.js";
+import { fetchWAQIData, fetchOpenWeatherAQIData, fetchWAQIDataLatLon } from "./api.js";
 import { getAQIClass } from "./helper.js";
 
 // Handle Tab switching
@@ -32,7 +32,7 @@ function setupTabs() {
     });
 }
 
-// For Home Tabs
+// For Home Tab
 
 // Update the home tab with AQI data
 async function updateHomeTab() {
@@ -58,4 +58,57 @@ function updateAQIDisplay(data) {
     $('.AQI-color').next("span").text(result.aqiText);
 }
 
-export {setupTabs, updateHomeTab};
+// For Live Tab
+function searchButtonLive() {
+    $('#search-live').on('click', async function () {
+        let selectedElement;
+
+        // Determine the selected element
+        if ($('#city-select-live').prop('selectedIndex') > 0) {
+            // If a city is selected
+            selectedElement = $('#city-select-live').find(':selected');
+        } else {
+            // Fallback to state if no city is selected
+            selectedElement = $('#state-select-live').find(':selected');
+        }
+
+        // Retrieve lat and lon from the selected element
+        const lat = selectedElement.data('lat');
+        const lon = selectedElement.data('lon');
+
+        if (!lat || !lon) {
+            alert("Latitude and longitude not available for the selected location.");
+            return;
+        }
+
+        // Fetch and update data using lat/lon
+        const waqiData = await fetchWAQIDataLatLon(lat, lon);
+        const owData = await fetchOpenWeatherAQIData(lat, lon, 'air_pollution');
+        const weatherData = await fetchOpenWeatherAQIData(lat, lon, 'weather');
+
+        const UiClass = getAQIClass(waqiData.data.aqi);
+
+        // Update UI and charts
+        const $liveAQIContainer = $('#live-aqi-colour').parent().parent();
+        $liveAQIContainer.removeClass(function (index, className) {
+            return (className.match(/(^|\s)bg\S+/g) || []).join(' ');
+        });
+        $liveAQIContainer.addClass("bg" + UiClass.aqiClass);
+
+        $('#live-aqi-colour').text(waqiData.data.aqi);
+        $('#live-aqi-colour').next("span").text(UiClass.aqiTitle);
+
+        // update temp, humidity, windspeed
+        const tempC = weatherData.main.temp - 273.15;
+        $('#live-temperature').text(tempC.toFixed(2) + "  °C");
+        $('#live-humidity').text(weatherData.main.humidity + "  %");
+        $('#live-windspeed').text(weatherData.wind.speed + "  m/s");
+
+        // update chart
+
+        const pollutantData = Object.values(owData.list[0].components);
+        updatePollutantLiveChart(pollutantData);
+    });
+}
+
+export {setupTabs, updateHomeTab, searchButtonLive};
